# jaxrs/src/main/resources/schema/swagger.yaml
openapi: 3.0.1
info:
  title: Fuse TryIt - Jaxrs API
  version: 1.0.0
...
paths:
  /user/{userId}:
    delete:
      summary: Delete the user with the given id
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "[No Content] Sucessfully delete the user in Keycloak."
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
  /user/{userId}/model/{modelId}:
    get:
      summary: Get the model for the given id.
      operationId: getModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Model for the given id could be accessed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
        "404":
          description: "[Not Found] The model for the given id could not be found."
    post:
      summary: Update the given model.
      operationId: updateModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserModel'
      responses:
        "200":
          description: "[OK] Successfully updated the model."
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
    delete:
      summary: Delete the model with the given id.
      operationId: deleteModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "[No Content] Successfully deleted the model."
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
  /user/{userId}/model/{modelId}/{runtime}/build:
    get:
      summary: Schedule the build process for the model.
      operationId: buildModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Successfully scheduled the model build."
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
        "404":
          description: "[Not Found] The model for the given id could not be found."
  /user/{userId}/model/{modelId}/{runtime}/download:
    get:
      summary: Download the build target
      operationId: downloadBuildTarget
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Found the requested build target."
          content:
            application/octet-stream: {}
        "401":
          description: "[Unauthorized] If the provided credentials were not valid."
        "404":
          description: "[Not Found] The target file was not found."
  /user/{userId}/model/{modelId}/{runtime}/status:
    get:
      summary: Get the current build status
      operationId: getBuildStatus
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Found the status for the requested project."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MavenBuildHandle'
        "401":
          description: "[Unauthorized] If the provided credentials were not valid."
        "404":
          description: "[Not Found] The project for the given id was not found."
  /user/{userId}/models:
    get:
      summary: Get the list of the given user's model ids.
      operationId: getModels
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] List of user model definitions."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModelList'
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
    put:
      summary: Create a new integration model
      operationId: createModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserModelAdd'
      responses:
        "201":
          description: "[Created] Model was successfully created."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
  /user/{userId}/state:
    get:
      summary: Fetch the current state for the given user id
      operationId: getUserState
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Found the requested user state."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserState'
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
...
components:
  schemas:
    MavenBuildHandle:
      required:
      - id
      - location
      - status
      type: object
      properties:
        id:
          type: string
        location:
          type: string
          format: uri
        status:
          type: string
          enum:
          - Scheduled
          - Running
          - Success
          - Failure
          - Removed
          - NotFound
    UserModel:
      required:
      - modelId
      - userId
      type: object
      properties:
        content:
          type: string
        modelId:
          type: string
        userId:
          type: string
    UserModelAdd:
      required:
      - content
      - userId
      type: object
      properties:
        content:
          type: string
        userId:
          type: string
    UserModelList:
      required:
      - userId
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/UserModel'
        userId:
          type: string
...
    UserState:
      required:
      - email
      - firstName
      - id
      - lastLogin
      - lastName
      - status
      - username
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        id:
          type: string
        lastLogin:
          type: string
          format: date-time
        lastName:
          type: string
        logins:
          type: integer
          format: int32
        status:
          type: string
        username:
          type: string
...